add custom metrics for Aurora
This commit adds a custom metric for detecting
when the Angular image directive is used and
whether it is marked as "priority". This was
previously difficult to detect using HTTP
Archive because most Angular apps are client-side
rendered, so the output would not be available
in the normal response body table. Closes #59.