add custom metrics for Aurora (#66)
* add custom metrics for Aurora

This commit adds a custom metric for detecting
when the Angular image directive is used and
whether it is marked as "priority". This was
previously difficult to detect using HTTP
Archive because most Angular apps are client-side
rendered, so the output would not be available
in the normal response body table. Closes #59.

* Update dist/aurora_apis.js

Co-authored-by: Rick Viscomi <rviscomi@users.noreply.github.com>

* update metric name

* linting

Co-authored-by: Rick Viscomi <rviscomi@users.noreply.github.com>