detect bound priority attr usages
Previously, the ng_priority_img_count custom metric
could only detect unbound usages of the "priority"
attribute because it was checking for it directly.
With this change, we check the "fetchpriority"
attribute instead because that will also catch
cases where the attribute was bound (in which case
it disappears from the DOM).